+++ test/integration/activerecord_backgrounding_test.rb
@@ -28,7 +28,7 @@ describe "ActiveRecord Backgrounding" do
       record = attacher.record.class.find(attacher.record.id)
 
       attacher.class
-        .retrieve(model: record, name: attacher.name, data: attacher.data)
+        .retrieve(model: record, name: attacher.name, file: attacher.file_data)
         .atomic_promote
     end
 
@@ -103,7 +103,7 @@ describe "ActiveRecord Backgrounding" do
         record = attacher.record.class.find(attacher.record.id)
 
         attacher.class
-          .retrieve(model: record, name: attacher.name, data: attacher.data)
+          .retrieve(model: record, name: attacher.name, file: attacher.file_data)
           .atomic_promote
       end
     end